+++ src/dynlib/steppers/__init__.py
@@ -1,5 +1,5 @@
 # src/dynlib/steppers/__init__.py
-from .base import StepperMeta, StepperInfo, StepperSpec
+from .base import StepperCaps, StepperMeta, StepperInfo, StepperSpec
 from .registry import register, get_stepper, registry
 
 # Import concrete steppers to trigger auto-registration
@@ -8,6 +8,6 @@ from .ode import euler, rk4, rk45
 from .ode import ab2, ab3
 
 __all__ = [
-    "StepperMeta", "StepperInfo", "StepperSpec",
+    "StepperCaps", "StepperMeta", "StepperInfo", "StepperSpec",
     "register", "get_stepper", "registry",
 ]